Updated Deck
The logic that determined what card should be made based on colour and type has become an independent private method. searchColourTypeCards was updated to searchSpecificCard with parameters String String, overloading searchSpecificCard(Card). The new search is public, and is used to get the index of the first instance of an object found; no longer returns the sum.