Updated Deck
The logic that determined what card should be made based on colour and type has become an independent private method, which considers capital letters for card types but also full name (capitalized first letter). searchColourTypeCards was updated to searchSpecificCard with parameters String String, overloading searchSpecificCard(Card). The new search is public, and is used to get the index of the first instance of an object found; no longer returns the sum.
Basically, you search for a card of similar properties and return it.