Fully Working Game Class and Simulation
All loops, logic, and effects are integrated. You can play a legitimate game. The CPU logic is sort of non random because their decks are sorted and they linear search for first instance,  meaning there is a pattern to the way they play. As well, WHERE ARE ALL THESE WILD CARDS COMING FROM?