WildCard Frenzy Fixed, Loop Back Not Fixed
The reason people had so many wild cards is because upon effect of a +4 or +2, the playermanager draws the cards from the discard pile instead of the draw pile. Additionally, added small change where upon moving a card to another deck, if not valid, returns the card back to the donor and REMOVES it from the reciever. Fixed GUI slightly.