Reveal Cards, Wait Method, Fixed Looping ApplyCard(previous)
You can now reveal cards of public piles or of player piles. Thread.sleep is now a method on its own in Uno, can be called anywhere. There was a bug which a player played an effect based card, like +2, if following players draw a card instead of playing them, the skip effect from +2 continues to apply. This should be fixed, not tested yet.