Beta 1.4
Combined Type and Ascending sort in Deck into one method for dual condition.
Discovered a bug where when two people in tutorial forget to call uno, the punishment only affects one.
Also discovered a bug where if a wild card was played first, the person deciding the colour is not the first person displayed, but the 2nd person. This is because of the initial sorting problem